# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Web-Engineering I Projekt 1 - REST documentation"
  },
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/data/umsaetze": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Umsatzinformationen",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Umsatz"
            }
          }
        }
      },
      "post": {
        "tags": [
          "data"
        ],
        "summary": "Kauf / Verkauf von Aktien",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AktieKaufVerkauf"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/AktieKaufVerkauf"
            }
          },
          "422": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/data/umsaetze/{umsatzID}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Finde Umsatz auf Basis der ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "umsatzID",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/UmsatzId"
            }
          },
          "422": {
            "description": "Unprocessable Entity / Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "AktieKaufVerkauf": {
      "type": "object",
      "properties": {
        "aktie": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "example": "Microsoft"
            }
          }
        },
        "anzahl": {
          "type": "number",
          "example": 2
        }
      }
    },
...
    "Umsatz": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "aktie": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Microsoft"
              },
              "preis": {
                "type": "number",
                "example": 91.01
              },
              "anzahlVerfuegbar": {
                "type": "number",
                "example": 460
              }
            }
          },
          "anzahl": {
            "type": "number",
            "example": 40
          }
        }
      }
    },
    "UmsatzId": {
      "type": "object",
      "properties": {
        "aktie": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "example": "Microsoft"
            },
            "preis": {
              "type": "number",
              "example": 91.01
            },
            "anzahlVerfuegbar": {
              "type": "number",
              "example": 460
            }
          }
        },
        "anzahl": {
          "type": "number",
          "example": 40
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "some error"
        }
      }
    }
  }
}